requirements.txt
tensorflow
flask
flask-cors
pillow
matplotlib
seaborn
scikit-learn
pandas
opencv-python
numpy